requirements-dev.txt
chalice
black
mypy
pytest
syrupy
py-pglite
psycopg
types-requests-oauthlib
apispec
apispec-pydantic-plugin
pre-commit
boto3-stubs
PyMuPDF
pymupdf-stubs
